Whoops. That's not how you set the # of threads. Forgot the variable in Kokkos_For is actually just the number of iterations. Also, got tired of cleaning in kokkos_gpu. Now we will only clean if I provided an argument.
"Another rebase."

+++ scripts/kokkos_gpu.sh
@@ -1,7 +1,9 @@
-if [ -f "CMakeCache.txt" ]; then
-  rm CMakeCache.txt
-fi
 
+if [ -z "$1" ]; then
+  if [ -f "CMakeCache.txt" ]; then
+    rm CMakeCache.txt
+  fi
+fi
 base_directory=$(basename "$PWD")
 
 if [[ "$base_directory" != "build" && "$base_directory" != "cmake-build-debug" ]]; then
